pyproject.toml
# (C) Copyright 2024 Anemoi contributors.
#
# This software is licensed under the terms of the Apache Licence Version 2.0
# which can be obtained at http://www.apache.org/licenses/LICENSE-2.0.
#
# In applying this licence, ECMWF does not waive the privileges and immunities
# granted to it by virtue of its status as an intergovernmental organisation
# nor does it submit to any jurisdiction.

[build-system]
requires = [ "setuptools>=60", "setuptools-scm>=8" ]

[project]
name = "anemoi-transform"

description = "A package to hold various data transformation functions to support training of ML models on ECMWF data."
keywords = [ "ai", "tools" ]

license = { file = "LICENSE" }
authors = [
  { name = "European Centre for Medium-Range Weather Forecasts (ECMWF)", email = "software.support@ecmwf.int" },
]

requires-python = ">=3.9"

classifiers = [
  "Development Status :: 4 - Beta",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: Apache Software License",
  "Operating System :: OS Independent",
  "Programming Language :: Python :: 3 :: Only",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Programming Language :: Python :: Implementation :: CPython",
  "Programming Language :: Python :: Implementation :: PyPy",
]

dynamic = [ "version" ]
dependencies = [
  "anemoi-utils>=0.4.11",
  "cfunits",
  "earthkit-data>=0.12",
  "earthkit-meteo>=0.3",
  "earthkit-regrid>=0.4",
]

optional-dependencies.all = [ "anemoi-transform" ]
optional-dependencies.dev = [ "anemoi-transform[all,docs,tests]" ]

optional-dependencies.docs = [
  "nbsphinx",
  "pandoc",
  "requests",
  "sphinx",
  "sphinx-argparse<0.5",
  "sphinx-rtd-theme",
  "termcolor",
]

optional-dependencies.tests = [ "pytest" ]

urls.Documentation = "https://anemoi-transform.readthedocs.io/"
urls.Homepage = "https://github.com/ecmwf/anemoi-transform/"
urls.Issues = "https://github.com/ecmwf/anemoi-transform/issues"
urls.Repository = "https://github.com/ecmwf/anemoi-transform/"

scripts.anemoi-transform = "anemoi.transform.__main__:main"

[tool.setuptools_scm]
version_file = "src/anemoi/transform/_version.py"

[tool.mypy]
strict = false
exclude = [
  "docs/**",
]

[tool.pydocstringformatter]
write = true
exclude = "docs/**"
style = "numpydoc"
